### 2021-06-10 #3342 @csoneson
👋🏻 @rpetit3, @ayush9pandey - would you be interested in reviewing this submission for JOSS?

> OSTIR: open source translation initiation rate prediction
