### 2021-06-11 #3342 @csoneson
👋🏻 @standage - would you be interested in reviewing this submision for JOSS?

> OSTIR: open source translation initiation rate prediction
