### 2021-06-11 #3362 @whedon
[REVIEW]: OSTIR: open source translation initiation rate prediction
**Submitting author:** @jeffreybarrick (<a href="http://orcid.org/0000-0003-0888-7358">Jeffrey Barrick</a>)
**Repository:** <a href="https://github.com/barricklab/ostir" target ="_blank">https://github.com/barricklab/ostir</a>
**Version:** v1.0.6
