### 2021-06-23 #3362 @csoneson
@standage Was this resolved by the reassignment by whedon above (just to check before investigating further)?
