### 2021-06-23 #3362 @standage
@csoneson Unfortunately no, I'm still unable to edit the checklist, either by clicking on the boxes or by editing the post/comment.
