### 2021-06-24 #3362 @arfon
I've just checked again that there's an invite and it's definitely showing up on GitHub https://github.com/openjournals/joss-reviews/invitations

@standage - perhaps try logging in to GitHub with a different browser before clicking the invite link? I'm not sure what's going on here sorry.
