### 2021-06-29 #3362 @ayush9pandey
I've been having issues with making the software work on my Windows machine due to a dependency (Vienna RNA) that I haven't been able to setup correctly. I have created an issue in their repository hoping for some help: https://github.com/ViennaRNA/ViennaRNA/issues/130 (posting here for tracking the review progress).
