### 2021-06-30 8cd7a8d
Removed reference to unix-only 'resource' package
