### 2021-06-30 9f545d6
Hardcoded output line endings
