### 2021-06-30 f7bb1cc
Wrap filepaths of test directories in quotes to resolve problematic spaces
