### 2021-07-05 #3362 @croots
Hi @ayush9pandey

Are you still having issues with ViennaRNA and OSTIR on Windows? We went back and noticed some issues with the unit testing on Windows with previous versions and make the necessary fixes in OSTIR v1.0.5 (it may take a bit for Bioconda to update to the latest version - install from source or pypi). You should be able to validate your ViennaRNA installation with `RNAfold --version`.
