### 2021-07-09 #3362 @standage
The authors describe OSTIR, a Python application for predicting translation initiation rates for bacterial mRNAs.
For each input sequence, OSTIR identifies putative start codons and reports the corresponding translation initiation rate and several other terms relating to the free energy of the putative binding event.

The calculations performed by OSTIR provide insight into a fundamental biochemical process, with a wide variety of applications in research and industry.
It is released under an open-source license that, unlike alternative software packages, permits its integration into custom workflows without burdensome licensing or user registration.
OSTIR appears to be well implemented and works as advertised in my hands.

- The recommended installation procedure was quick, seamless, and follows broadly adopted community standards and best practices.
- OSTIR's command line interface is self-documenting, and displays a detailed usage statement when the `ostir` command is invoked with no arguments or with the `-h` or `--help` arguments.
- The OSTIR source code distribution includes an automated test suite. Since I installed OSTIR using a package manager rather than from source, I did not run the complete test suite. But several examples were provided in the documentation and source distribution, and these were easy to test and confirm.
- OSTIR performed well for a variety of tests designed to stress or break the code. In each of the following cases, OSTIR printed a helpful error message and terminated gracefully—no crashes, stack traces, or cryptic messages.
	- protein sequence as input
	- random text file as input
	- single nucleotide sequence as input

This tool looks like a valuable contribution to the field. However, I have a concern about documentation and a recommendation about testing.

## Concern: guidance for evaluating output

In the paper summary, the authors claim—rightly so in my opinion—that the capabilities provided by OSTIR are of potential interest to a broad community and have applications in many areas of research and industry. However, the software documentation is very minimal and suitable only for a very narrow audience; that is, folks well-versed in biochemistry and models of RNA thermodynamics. For e.g. a newcomer to the field or a potentially interested user from an adjacent field, the documentation provides little to help the user interpret and evaluate the results of OSTIR output.

To be clear, a comprehensive exposition of RNA thermodynamics is out of scope for this submission, and a rigorous evaluation of the corresponding models is out of scope for this review. But if OSTIR is to be truly accessible to the broad audience the authors describe in their paper, it's not unreasonable to expect that the authors provide at least minimal guidance for interpreting the output of the software. Doing so would certainly be in the spirit of the open source community ethos embraced by JOSS.

As part of my evaluation of OSTIR, I grabbed a handful of bacterial mRNA sequences (coding sequence + some flanking sequence to approximate UTRs on both sides) from published genomes and compared the OSTIR output to output for random sequences of 100 `[ACGT]` characters. There is little in the documentation that would help me answer the following questions.

- What should one expect from such an exercise?
- I observed some notable differences in predicted translation rate and ΔG: how should I interpret those differences?
- Depending on the size and composition of the mRNA, there may be numerous putative start codons. How does one distinguish true initiation sites from spurious predictions?

The answers to these questions are no doubt obvious to the authors—and to a narrow audience with an intuitive understanding of RNA thermodynamics—but not as obvious to many in the broader genome informatics and bioinformatics communities. Providing some basic guidance for interpreting the OSTIR output would be a valuable improvement to this submission.

## Recommendation: configure a continuous integration environment

The OSTIR source code distribution includes an automated test suite that is presumably run routinely by the authors. I'd recommend that the authors configure a continuous integration (CI) environment that automatically executes the test suite on a frequent basis.

CI can be configured to:

1. run on a schedule (i.e. a monthly cron job); or
2. run on an event trigger (e.g. whenever the repository is updated); or
3. both

However the CI is configured, each CI build starts with a fresh VM and builds the environment from scratch using instructions provided by the authors. This has numerous benefits.

- Committing setup instructions to a CI configuration file mitigates the all-to-common "Well, it works on *my* machine!" issues.
- While a project is under active development, event triggers ensure the test suite is run often and that any new bugs are identified early.
- When a project is in maintenance mode, scheduled builds ensure that new issues (such as incompatibilities with newer versions of a package dependency) are caught early, while they are still easy to manage. Allowing such incompatibilities to accumulate over time severely limits the viable lifetime of a software project.
- If and when members of the community submit bug fixes, suggested features, or other patches/pull requests, the CI build status (pass, fail, error) provides valuable feedback both to the OSTIR authors and the pull request author.

Numerous CI providers are available for free to open source software projects: GitHub Actions is an obvious choice built right into the GitHub platform, but third-party providers such as Drone CI and Circle CI are also available.
