### 2021-07-12 5029d63
Automated tests on linux/mac with Github Actions
