+++ environment.yml
@@ -1,0 +1,30 @@
+name: ostir_env
+channels:
+  - conda-forge
+  - bioconda
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=conda_forge
+  - _openmp_mutex=4.5=1_gnu
+  - ca-certificates=2021.5.30=ha878542_0
+  - certifi=2021.5.30=py39hf3d152e_0
+  - ld_impl_linux-64=2.36.1=hea4e1c9_1
+  - libffi=3.3=h58526e2_2
+  - libgcc-ng=9.3.0=h2828fa1_19
+  - libgomp=9.3.0=h2828fa1_19
+  - libstdcxx-ng=9.3.0=h6de172a_19
+  - ncurses=6.2=h58526e2_4
+  - openssl=1.1.1k=h7f98852_0
+  - pip=21.1.3=pyhd8ed1ab_0
+  - python=3.9.6=h49503c6_0_cpython
+  - python_abi=3.9=2_cp39
+  - readline=8.1=h46c0cb4_0
+  - setuptools=49.6.0=py39hf3d152e_3
+  - sqlite=3.36.0=h9cd32fc_0
+  - tk=8.6.10=h21135ba_1
+  - tzdata=2021a=he74cb21_1
+  - viennarna=2.4.18=py39h98c8e45_0
+  - wheel=0.36.2=pyhd3deb0d_0
+  - xz=5.2.5=h516909a_1
+  - zlib=1.2.11=h516909a_1010
+prefix: /home/croots/miniconda3/envs/ostir_env
