### 2021-07-12 a617985
Updates environment.yml and package_and_test.yml
