### 2021-07-15 #3362 @ayush9pandey
The authors present an open source software package to predict translation initiation rates of mRNAs in bacteria. The primary goal of this package is to provide an open source and command line tool that can be used for translation rate prediction since other tools with similar features are not open source yet. In particular, the widely used software RBS calculator is proprietary and can only be used through a webpage. Towards that end, OSTIR opens up new and interesting possibilities as it can become a part of a bigger pipeline to analyze/design biological systems.

## Some comments:

1. There are two suggested approaches to install OSTIR - one using `bioconda` and another using `PyPi` distribution. The installation using `bioconda` does not require the user to install `ViennaRNA` (an important OSTIR dependency) manually. However, to use the PyPi distribution of OSTIR, the user needs to install ViennaRNA before installing OSTIR. For Windows machines, I was not able to get ViennaRNA to run successfully and had to switch to a linux machine to get it all running. So, I would suggest that the authors recommend the bioconda installation method in their README for the package.
2. The GitHub does not provide a contribution guidelines document/page and the issues page could include a few labeled topics (such as `help-wanted`, `enhancement`, etc.) that may help a new open source developer contribute to the project.
3. For a scientist who is a potential user of OSTIR, the README documentation is a good start but it could be improved with a dedicated documentation page (either using GitHub wiki or another open source documentation web service).
4. A detailed API documentation would also help any developers include OSTIR as a part of their software pipelines. Since a major contribution of OSTIR is that it provides an open source code that can predict translation rate initiation given the mRNA sequences, I think various tools could use the OSTIR API. For example, a software designed for forward design of synthetic biological circuits could massively benefit from OSTIR outputs.
