+++ README.md
@@ -1,5 +1,5 @@
 # OSTIR (Open Source Translation Initiation Rates)
-[![install with bioconda](https://img.shields.io/badge/install%20with-bioconda-brightgreen.svg?style=flat)](http://bioconda.github.io/recipes/ostir/README.html) ![https://github.com/barricklab/ostir/actions/workflows/package_and_test.yml](https://github.com/barricklab/ostir/actions/workflows/package_and_test.yml/badge.svg)
+[![install with bioconda](https://img.shields.io/badge/install%20with-bioconda-brightgreen.svg?style=flat)](http://bioconda.github.io/recipes/ostir/README.html) [![Status](https://github.com/barricklab/ostir/actions/workflows/package_and_test.yml/badge.svg)](https://github.com/barricklab/ostir/actions/workflows/package_and_test.yml)
 
 
 `OSTIR` is a
