### 2021-08-09 112d428
Added E. coli scripts and images for Wiki
