+++ ecoli/ostir_on_ecoli.py
@@ -1,0 +1,182 @@
+from Bio import SeqIO
+from ostir.ostir import run_ostir, save_to_csv
+import sys
+from os.path import exists
+import os
+import csv
+from datetime import datetime
+
+def ostir_from_gb(gb):
+    genbank = SeqIO.read(gb, "genbank")
+    sequence = str(genbank.seq)
+    length = len(sequence)
+    strand = "-"
+    bp_to_iter = 50000
+    start_pos = 0
+    end_pos = start_pos + bp_to_iter
+    print(f'{length} bases to go through')
+    while True:
+        outpath = f"ecoli/{start_pos}_{end_pos}_{strand}.csv"
+        if not exists(outpath):
+            print(f'Running ostir on E. coli {start_pos} to {end_pos} on {strand} strand')
+            results = run_ostir(in_seq=sequence, start=start_pos, end=end_pos, name=None, aSD=None, threads=8, decimal_places=4,
+                                verbose=False)
+            column_names = ['name', 'start_codon', 'start_position', 'expression', 'RBS_distance_bp', 'dG_total',
+                            'dG_rRNA:mRNA', 'dG_mRNA', 'dG_spacing', 'dG_standby', 'dG_start_codon']
+            save_to_csv(column_names, results, outpath)
+            print(f'[{datetime.now()}] Results written to {outpath}', file=sys.stderr)
+
+        if end_pos == length:
+            if strand == "-":
+                break
+            else:
+                strand = "-"
+                start_pos = 0
+                end_pos = start_pos + bp_to_iter
+                sequence = str(genbank.seq.reverse_complement())
+                continue
+
+        start_pos = end_pos+1
+        end_pos = end_pos + bp_to_iter if (end_pos + bp_to_iter) <= length else length
+    print('Finished OSTIR on the genome')
+
+def start_codon_csv(gb):
+    genbank = SeqIO.read(gb, "genbank")
+
+    forward_CDS = {}
+    reverse_CDS = {}
+    forward_starts_ostir = []
+    reverse_starts_ostir = []
+
+    print('Building dataframe for start codons')
+    print(f'{len(os.listdir("ecoli"))} files to go through')
+
+
+    for filenumber, file in enumerate(os.listdir('ecoli')):  # Get start positions from CSVs
+        print(f'Fetching start for file {filenumber+1} of {len(os.listdir("ecoli"))}')
+        with open(f'ecoli/{file}') as infile:
+            csv_file = csv.reader(infile)
+            if '+' in file:
+                target_list = forward_starts_ostir
+                strand = '+'
+            elif '-' in file:
+                target_list = reverse_starts_ostir
+                strand = '-'
+            else:
+                continue
+            for i, row in enumerate(csv_file):
+                if i == 0:
+                    continue
+                data = [int(row[2]),  # Start position
+                        float(row[3]),  # Expression estimate
+                        strand]
+                target_list.append(data)
+
+    gb_length = len(str(genbank.seq))
+
+    reverse_starts_ostir = [[gb_length-start_position[0], start_position[1], start_position[2]] for
+                            start_position in reverse_starts_ostir]
+
+    print("Parsing genbank file")
+    for record in SeqIO.parse(gb, 'genbank'):  # Get CDS positions from gb file
+        for feature in record.features:
+            if feature.type == "CDS":
+                data = [int(feature.location.start), int(feature.location.end), feature.qualifiers["gene"]]
+                if feature.strand == 1:
+                    target_dict = forward_CDS
+                    target_dict[str(int(feature.location.start))] = data
+                elif feature.strand == -1:
+                    target_dict = reverse_CDS
+                    target_dict[str(int(feature.location.end))] = data
+                else:
+                    raise ValueError('Found tripple stranded DNA?')
+
+
+    length_forward = len(forward_starts_ostir)
[…]
