+++ README.md
@@ -1,12 +1,11 @@
 # OSTIR (Open Source Translation Initiation Rates)
-[![install with bioconda](https://img.shields.io/badge/install%20with-bioconda-brightgreen.svg?style=flat)](http://bioconda.github.io/recipes/ostir/README.html) [![Status](https://github.com/barricklab/ostir/actions/workflows/package_and_test.yml/badge.svg)](https://github.com/barricklab/ostir/actions/workflows/package_and_test.yml)
+[![Status](https://github.com/barricklab/ostir/actions/workflows/package_and_test.yml/badge.svg)](https://github.com/barricklab/ostir/actions/workflows/package_and_test.yml)
 
 
 `OSTIR` is a
 Python package for predicting the rates at which ribosomes will bind to and initiate
-translation from different start codons in bacterial mRNAs. It uses the ViennaRNA software
-suite to perform the necessary free energy calculations. The code builds on the last open
-source version of the
+translation from different start codons in bacterial mRNAs. It uses the ViennaRNA Package to perform 
+the necessary free energy calculations. The code builds on the last open source version of the
 [RBS calculator](https://github.com/hsalis/Ribosome-Binding-Site-Calculator-v1.0)
 that implements the calculations described in [Salis 2009](https://doi.org/10.1038/nbt.1568).
 
@@ -14,6 +13,10 @@ that implements the calculations described in [Salis 2009](https://doi.org/10.10
 input with command line parameters or CSV input that can define
 parameters on a per-sequence basis. Additionally, `OSTIR` supports multi-threaded
 execution, accelerating use cases that require the analysis of very large sequences.
+
+### [Please see the OSTIR Wiki for full documentation](https://github.com/barricklab/ostir/wiki)
+
+# Quickstart
 
 ## Installation
 [![install with bioconda](https://img.shields.io/badge/install%20with-bioconda-brightgreen.svg?style=flat)](http://bioconda.github.io/recipes/ostir/README.html)
@@ -27,130 +30,38 @@ From Pip (for experts; Linux, macOS, Windows):
 - Download and install ViennaRNA, following the instructions [here](https://www.tbi.univie.ac.at/RNA/).
 - Run `pip install ostir`
 
-From Source (for developers; Linux, macOS, Windows):
-- Clone this repository
-- Navigate to the `./ostir/` directory
-- Install/update the necessary packaging tools with `python3 -m pip install --user --upgrade setuptools wheel`
-- Package the code with `python3 setup.py sdist bdist_wheel`
-- Install the package with `python3 -m pip install ./`
-- Download and install ViennaRNA, following the instructions [here](https://www.tbi.univie.ac.at/RNA/).
-- To test your install run `python -m unittest`
+For information on installing for development see the [Wiki Documentation](https://github.com/barricklab/ostir/wiki/Installation).
 
 ## Command Line Usage
 
-OSTIR can be executed via the included command-line script, `ostir`.
-
+Print OSTIR help:
 ```
-usage: ostir [-h] -i str/filepath [-o filepath] [-v] [-s int] [-e int] [-a str] [-p] [-q] [-j int] [-t [string|csv|fasta]]
-
-Open Source Transcription Initiation Rates
-
-optional arguments:
-  -h, --help            show this help message and exit
-  -i str/filepath, --input str/filepath
-                        Input filename (FASTA/CSV) or DNA/RNA sequence. For CSV input files, there must be a 'seq' or 'sequence'
-                        column. Other columns will override any options provided at the command line if they are present:
-                        'name/id', 'start', 'end', 'anti-Shine-Dalgarno'.
-  -o filepath, --output filepath
[…]
