### 2021-08-15 #3362 @jeffreybarrick
Thank you @standage and @ayush9pandey for the helpful feedback! And thanks @csoneson for organizing the review process.

Here's a run-down of the updates we made to improve OSTIR in response to each of your comments. Let us know if we missed anything or if you have additional comments/suggestions.

As part of this effort, we re-organized the online documentation, which had been a single `README.md`, into a [set of GitHub wiki pages](https://github.com/barricklab/ostir/wiki). The main [`README.md`](https://github.com/barricklab/ostir/blob/master/README.md) is now a shorter Quickstart guide that links to the more detailed documentation on the Wiki.

# Reviewer 1 (@standage)

## Concern: guidance for evaluating output

Great suggestion. We added a [Background Page](https://github.com/barricklab/ostir/wiki/Background) to the Wiki that has the following information:
* A short overview of how gene expression works in bacteria.
* An overall description of the thermodynamic model that defines each of the energy (&Delta;G) terms reported by OSTIR.
* Plots showing the results of running OSTIR on the entire _E. coli_ genome.

The last bullet point is especially relevant to the reviewer's questions about interpreting the predicted expression values. It shows that a true start codon (one that begins an annotated gene) gives a prediction of roughly 1000 on average. Most "random" potential start codons elsewhere in the genome give scores below 10.

## Recommendation: configure a continuous integration environment

Thanks for the suggestion to get this set up at the outset to make long-term maintenance easier.

We set up GitHub Actions for continuous integration. The [build and test action](https://github.com/barricklab/ostir/actions/workflows/package_and_test.yml) performs linting and Linux/MacOSX installation and consistency tests. We have it configured as both a triggered and scheduled (monthly) task and added a corresponding passing/failed badge to the [main OSTIR GitHub page](https://github.com/barricklab/ostir).

# Reviewer 2 (@ayush9pandey)

## 1. Installation instructions/issues

We updated the [installation instructions](https://github.com/barricklab/ostir/wiki/Installation) as suggested to recommend using `bioconda`. However, there isn't a Windows `conda` package for the ViennaRNA requirement, so we also kept the previous `pypi` directions for a Windows install with a proviso that it is an 'expert option'.

We're not sure about your difficulty installing on Windows. Our best guess is that the installation issue was a `$PATH` setup issue where the terminal was not finding the ViennaRNA executables on the reviewer's machine as described here ViennaRNA/ViennaRNA#130.

One of us (@croots) is a Windows user and has tried hard to maintain compatibility and consistency on this platform throughout OSTIR development, including by submitting issues to the ViennaRNA team when we found problems with earlier versions of ViennaRNA on Windows. We hope to maintain Windows functionality going forward, but recognize that this will be more of a challenge than it is for Linux/MacOS.

## 2. Labeled topics for encouraging contributions

Great idea. We added a couple of ideas for `enhancements` and a `help-wanted` item to kick things off on the [GitHub Issues Tracker for the project](https://github.com/barricklab/ostir/issues). We also added a short [Developer Information Wiki page](https://github.com/barricklab/ostir/wiki/Development) that includes a code of conduct.

## 3. Dedicated documentation

We expanded the documentation into a [set of GitHub wiki pages](https://github.com/barricklab/ostir/wiki) and changed the [`README.md`](https://github.com/barricklab/ostir/blob/master/README.md) to a Quickstart guide. Much of the content here is the same as it was in the original `README.md` (just re-organized into different pages), but we did add entirely new [Background](https://github.com/barricklab/ostir/wiki/Background) and [Developer Information](https://github.com/barricklab/ostir/wiki/Development) pages.

## 4. API Documentation

OSTIR exposes the Python function `run_ostir()` as an entry point as described on the updated [Python Module Usage Wiki page](https://github.com/barricklab/ostir/wiki/Python-Module-Usage). At this point, we think this is the full extent of an API that will be useful to users/developers of OSTIR.
