### 2021-08-17 #3362 @standage
Thanks for these responses @jeffreybarrick. The new background page in the wiki is indeed insightful, and the CI builds will hopefully ensure that OSTIR remains a valuable contribution to the community for many years requiring minimal maintenance.

@csoneson, I have no further concerns and recommend this submission for publication.
