### 2021-08-18 #3362 @ayush9pandey
Thanks for all the updates to the software @jeffreybarrick!
@csoneson I have completed all checkpoints for my review and can recommend this software for publication in JOSS.
