### 2021-08-20 #3362 @csoneson
👋🏻 @jeffreybarrick - I have looked through the submission and overall this looks good to me. A couple of things I noticed in the paper:

- [ ] the reference to Figure 1 (line 44) should be in parentheses
- [ ] "The Bioconda Team" should be spelled out in Grüning et al (2018)
- [ ] Several of the references lack issue and page numbers

Could you please fix these details (you can then ask whedon to generate a new pdf proof)? The next steps would then be to:

- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
