+++ JOSS/paper.bib
@@ -6,6 +6,7 @@
   month        = sep,
   journal      = {Nature Biotechnology},
   volume       = 27,
+  pages        = {946--950},
   doi          = {10.1038/nbt.1568},
   url          = {https://www.nature.com/articles/nbt.1568}
 }
@@ -37,6 +38,7 @@
   month        = oct,
   journal      = {ACS Synthetic Biology},
   volume       = 9,
+  pages        = {3145--3156},
   doi          = {10.1021/acssynbio.0c00394},
   url          = {https://pubs.acs.org/doi/abs/10.1021/acssynbio.0c00394}
 }
@@ -47,6 +49,7 @@
   month        = oct,
   journal      = {Bioinformatics},
   volume       = 26,
+  pages        = {2633--2634},
   doi          = {10.1093/bioinformatics/btq458},
   url          = {https://academic.oup.com/bioinformatics/article/26/20/2633/193535}
 }
@@ -57,6 +60,7 @@
   month        = nov,
   journal      = {Nature Methods},
   volume       = 6,
+  pages        = {26+},
   doi          = {10.1186/1748-7188-6-26},
   url          = {https://almob.biomedcentral.com/articles/10.1186/1748-7188-6-26}
 }
@@ -67,16 +71,18 @@
   month        = jan,
   journal      = {Journal of Computational Chemistry},
   volume       = 32,
+  pages        = {170--173},
   doi          = {10.1002/jcc.21596},
   url          = {https://onlinelibrary.wiley.com/doi/full/10.1002/jcc.21596}
 }
-@article{Zadeh:2018,
+@article{Grüning:2018,
   title        = {{Bioconda: sustainable and comprehensive software distribution for the life sciences}},
-  author       = {{Grüning}, B. and {Dale}, R. and {Sjödin}, A and {Chapman}, B. and {Rowe}, J. and {Tomkins-Tinch}, CH. and {Valieris}, R. and {Köster}, J. and The Bioconda Team},
+  author       = {{Grüning}, B. and {Dale}, R. and {Sjödin}, A and {Chapman}, B. and {Rowe}, J. and {Tomkins-Tinch}, CH. and {Valieris}, R. and {Köster}, J. and {{The Bioconda Team}},
   year         = 2018,
   month        = jul,
   journal      = {Nature Methods},
   volume       = 15,
+  pages        = {475--476},
   doi          = {10.1038/s41592-018-0046-7},
   url          = {https://www.nature.com/articles/s41592-018-0046-7}
 }
