+++ JOSS/paper.bib
@@ -77,7 +77,7 @@
 }
 @article{Grüning:2018,
   title        = {{Bioconda: sustainable and comprehensive software distribution for the life sciences}},
-  author       = {{Grüning}, B. and {Dale}, R. and {Sjödin}, A and {Chapman}, B. and {Rowe}, J. and {Tomkins-Tinch}, CH. and {Valieris}, R. and {Köster}, J. and {{The Bioconda Team}},
+  author       = {{Grüning}, B. and {Dale}, R. and {Sjödin}, A and {Chapman}, B. and {Rowe}, J. and {Tomkins-Tinch}, CH. and {Valieris}, R. and {Köster}, J. and {{The Bioconda Team}}},
   year         = 2018,
   month        = jul,
   journal      = {Nature Methods},
