+++ JOSS/paper.bib
@@ -17,7 +17,8 @@
   journal      = {Methods in Enzymology},
   volume       = 498,
   pages        = {19--42},
-  doi          = {10.1016/B978-0-12-385120-8.00002-4}
+  doi          = {10.1016/B978-0-12-385120-8.00002-4},
+  url          = {https://www.sciencedirect.com/science/article/abs/pii/B9780123851208000024}
 }
 @article{EspahBorujeni:2014,
   title        = {{Translation rate is controlled by coupled trade-offs between site accessibility, selective RNA unfolding and sliding at upstream standby sites}},
@@ -30,6 +31,7 @@
   doi          = {10.1093/nar/gkt1139},
   keywords     = {10,1093,2013,2014,2659 nucleic acids research,4,42,6,by coupled trade-offs,doi,gkt1139,nar,no,published online 14 november,translation rate is controlled,vol},
   pmid         = 24234441
+  url          = {https://academic.oup.com/nar/article/42/4/2646/2435298}
 }
 @article{Reis:2020,
   title        = {{An Automated Model Test System for Systematic Development and Improvement of Gene Expression Models}},
@@ -38,6 +40,7 @@
   month        = oct,
   journal      = {ACS Synthetic Biology},
   volume       = 9,
+  number       = 11,
   pages        = {3145--3156},
   doi          = {10.1021/acssynbio.0c00394},
   url          = {https://pubs.acs.org/doi/abs/10.1021/acssynbio.0c00394}
@@ -49,6 +52,7 @@
   month        = oct,
   journal      = {Bioinformatics},
   volume       = 26,
+  number       = 20,
   pages        = {2633--2634},
   doi          = {10.1093/bioinformatics/btq458},
   url          = {https://academic.oup.com/bioinformatics/article/26/20/2633/193535}
@@ -60,7 +64,7 @@
   month        = nov,
   journal      = {Nature Methods},
   volume       = 6,
-  pages        = {26+},
+  pages        = 26,
   doi          = {10.1186/1748-7188-6-26},
   url          = {https://almob.biomedcentral.com/articles/10.1186/1748-7188-6-26}
 }
