+++ JOSS/paper.bib
@@ -30,7 +30,7 @@
   pages        = {2646--2659},
   doi          = {10.1093/nar/gkt1139},
   keywords     = {10,1093,2013,2014,2659 nucleic acids research,4,42,6,by coupled trade-offs,doi,gkt1139,nar,no,published online 14 november,translation rate is controlled,vol},
-  pmid         = 24234441
+  pmid         = 24234441,
   url          = {https://academic.oup.com/nar/article/42/4/2646/2435298}
 }
 @article{Reis:2020,
