+++ JOSS/paper.bib
@@ -75,6 +75,7 @@
   month        = jan,
   journal      = {Journal of Computational Chemistry},
   volume       = 32,
+  number       = 1,
   pages        = {170--173},
   doi          = {10.1002/jcc.21596},
   url          = {https://onlinelibrary.wiley.com/doi/full/10.1002/jcc.21596}
