+++ JOSS/paper.bib
@@ -62,7 +62,7 @@
   author       = {{Lorenz}, R. and {Bernhart}, SH. and {Höner zu Siederdissen}, C. and {Tafer}, H. and {Flamm}, C. and {Stadler}, PF. and {Hofacker}, IL.},
   year         = 2011,
   month        = nov,
-  journal      = {Nature Methods},
+  journal      = {Algorithms for Molecular Biology},
   volume       = 6,
   pages        = 26,
   doi          = {10.1186/1748-7188-6-26},
