+++ JOSS/paper.bib
@@ -6,6 +6,7 @@
   month        = sep,
   journal      = {Nature Biotechnology},
   volume       = 27,
+  number       = 10,
   pages        = {946--950},
   doi          = {10.1038/nbt.1568},
   url          = {https://www.nature.com/articles/nbt.1568}
@@ -34,7 +35,7 @@
   url          = {https://academic.oup.com/nar/article/42/4/2646/2435298}
 }
 @article{Reis:2020,
-  title        = {{An Automated Model Test System for Systematic Development and Improvement of Gene Expression Models}},
+  title        = {{An automated model test system for systematic development and improvement of gene expression models}},
   author       = {{Reis}, AC. and {Salis}, HM.},
   year         = 2020,
   month        = oct,
@@ -87,6 +88,7 @@
   month        = jul,
   journal      = {Nature Methods},
   volume       = 15,
+  number       = 7,
   pages        = {475--476},
   doi          = {10.1038/s41592-018-0046-7},
   url          = {https://www.nature.com/articles/s41592-018-0046-7}
