### 2021-08-20 #3362 @jeffreybarrick
We think we've fixed all of these issues in the paper:
* [X]  the reference to Figure 1 (line 44) should be in parentheses
* [X]  "The Bioconda Team" should be spelled out in Grüning et al (2018)
* [X]  Several of the references lack issue and page numbers
