### 2021-08-20 26f6bcb
Update version number to 1.0.6
