### 2021-08-20 #3362 @jeffreybarrick
* [X]  Make a tagged release of the software, and list the version tag of the archived version here.
* [X]  Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.

**Version tag for archived version: v.1.0.6**
https://github.com/barricklab/ostir/releases/tag/v1.0.6

**DOI for Archived Version: 10.5281/zenodo.5227845**
https://zenodo.org/record/5227845#.YSABb9NKjm0

Thanks @csoneson for walking us through the steps! Let us know if there's anything else to correct.
